Improve prediction algorithm
Previously the prediction of an entire sentence was not found reliably. To combat this, add 'bootstrap' mechanism that makes multiple predictions and then heuristically create a multinomial distribution to pick the final prediction from. This does NOT produce a result 100% of the time but its way better and tweaking the parameters on the fly should allow for results very close to 100% of the time at the cost of a little more time for predictions.